Updated data merges to a single alpha7 gate to keep things clean for the final product.
