+++ src/M3Undle.Web/Program.cs
@@ -410,9 +410,6 @@ List<string> appliedMigrations;
 using (var scope = app.Services.CreateScope())
 {
     var db = scope.ServiceProvider.GetRequiredService<ApplicationDbContext>();
-    await RepairAlpha4MigrationHistoryAsync(db);
-    await StartupMigrationRepair.RepairAlpha5PartialSchemaAsync(db);
-    await RepairAlpha6SchemaAsync(db);
     appliedMigrations = db.Database.GetPendingMigrations().ToList();
     db.Database.Migrate();
     db.Database.ExecuteSqlRaw("PRAGMA journal_mode=WAL;");
@@ -717,174 +714,6 @@ static void EnsureWebRootExists()
             // a real error if no usable web root can be established.
         }
     }
-}
-
-// Repairs __EFMigrationsHistory when upgrading from alpha.4 databases that had four individual
-// migrations before they were consolidated into a single Alpha4_Schema migration.
-static async Task RepairAlpha4MigrationHistoryAsync(ApplicationDbContext db)
-{
-    var conn = db.Database.GetDbConnection();
-    if (conn.State != System.Data.ConnectionState.Open)
-        await conn.OpenAsync();
-
-    await using var checkTable = conn.CreateCommand();
-    checkTable.CommandText =
-        "SELECT COUNT(*) FROM sqlite_master WHERE type='table' AND name='__EFMigrationsHistory'";
-    if ((long)(await checkTable.ExecuteScalarAsync())! == 0)
-        return;
-
-    await using var check = conn.CreateCommand();
-    check.CommandText =
-        "SELECT COUNT(*) FROM \"__EFMigrationsHistory\" WHERE \"MigrationId\" IN (" +
-        "'20260314145015_Alpha4_StreamingSettings'," +
-        "'20260314152455_Alpha4_ProviderStreamLimits'," +
-        "'20260316121421_Alpha4_EpgSources'," +
-        "'20260317120000_Alpha4_StreamSettingsUi')";
-    if ((long)(await check.ExecuteScalarAsync())! == 0)
-        return;
-
-    await using var tx = await conn.BeginTransactionAsync();
-
-    await using var del = conn.CreateCommand();
-    del.Transaction = tx;
-    del.CommandText =
-        "DELETE FROM \"__EFMigrationsHistory\" WHERE \"MigrationId\" IN (" +
-        "'20260314145015_Alpha4_StreamingSettings'," +
-        "'20260314152455_Alpha4_ProviderStreamLimits'," +
-        "'20260316121421_Alpha4_EpgSources'," +
-        "'20260317120000_Alpha4_StreamSettingsUi')";
-    await del.ExecuteNonQueryAsync();
-
-    await using var ins = conn.CreateCommand();
-    ins.Transaction = tx;
-    ins.CommandText =
-        "INSERT OR IGNORE INTO \"__EFMigrationsHistory\" (\"MigrationId\", \"ProductVersion\") " +
-        "VALUES (@id, @ver)";
-    var pId = ins.CreateParameter();
-    pId.ParameterName = "@id";
-    pId.Value = "20260314145015_Alpha4_Schema";
-    ins.Parameters.Add(pId);
-    var pVer = ins.CreateParameter();
-    pVer.ParameterName = "@ver";
-    pVer.Value = "10.0.0";
-    ins.Parameters.Add(pVer);
-    await ins.ExecuteNonQueryAsync();
-
-    await tx.CommitAsync();
-}
-
-// Repairs databases that were upgraded through alpha.5 with a providers table that had
-// is_active removed outside of the normal migration path. Alpha6_ActiveProfile moves
-// is_active from providers to profiles, but its data-migration SQL references
-// providers.is_active which is already absent on these DBs, causing Migrate() to fail
-// with "no such column: p.is_active" and leaving profiles.is_active unset.
-//
-// When detected, this function manually applies the profiles-side schema change,
-// seeds a sensible active profile, and records Alpha6_ActiveProfile as applied so
-// Migrate() skips its broken data-migration step.
-static async Task RepairAlpha6SchemaAsync(ApplicationDbContext db)
-{
-    const string migrationId = "20260404000000_Alpha6_ActiveProfile";
-
-    var conn = db.Database.GetDbConnection();
-    if (conn.State != System.Data.ConnectionState.Open)
-        await conn.OpenAsync();
-
-    // Bail on a fresh DB — Migrate() will create everything from scratch.
-    await using var checkHistory = conn.CreateCommand();
-    checkHistory.CommandText =
-        "SELECT COUNT(*) FROM sqlite_master WHERE type='table' AND name='__EFMigrationsHistory'";
-    if ((long)(await checkHistory.ExecuteScalarAsync())! == 0)
-        return;
-
-    // Already applied — nothing to repair.
-    await using var checkApplied = conn.CreateCommand();
-    checkApplied.CommandText =
-        "SELECT COUNT(*) FROM \"__EFMigrationsHistory\" WHERE \"MigrationId\" = @id";
-    var p = checkApplied.CreateParameter();
-    p.ParameterName = "@id";
-    p.Value = migrationId;
-    checkApplied.Parameters.Add(p);
-    if ((long)(await checkApplied.ExecuteScalarAsync())! > 0)
-        return;
-
-    // No providers table yet — Migrate() will build the full schema.
-    await using var checkProviders = conn.CreateCommand();
-    checkProviders.CommandText =
-        "SELECT COUNT(*) FROM sqlite_master WHERE type='table' AND name='providers'";
-    if ((long)(await checkProviders.ExecuteScalarAsync())! == 0)
-        return;
-
-    // If providers.is_active exists the normal Alpha6 migration will run fine.
-    await using var checkProviderCol = conn.CreateCommand();
-    checkProviderCol.CommandText =
-        "SELECT COUNT(*) FROM pragma_table_info('providers') WHERE name='is_active'";
-    if ((long)(await checkProviderCol.ExecuteScalarAsync())! > 0)
-        return;
-
-    // Broken state confirmed: providers.is_active is missing and Alpha6 has never been
-    // recorded. Manually apply the profiles-side changes and mark the migration applied.
-    await using var checkProfileCol = conn.CreateCommand();
-    checkProfileCol.CommandText =
-        "SELECT COUNT(*) FROM pragma_table_info('profiles') WHERE name='is_active'";
-    var profileColExists = (long)(await checkProfileCol.ExecuteScalarAsync())! > 0;
-
-    await using var tx = await conn.BeginTransactionAsync();
-
-    if (!profileColExists)
-    {
-        await using var addCol = conn.CreateCommand();
-        addCol.Transaction = tx;
-        addCol.CommandText =
-            "ALTER TABLE \"profiles\" ADD COLUMN \"is_active\" INTEGER NOT NULL DEFAULT 0";
-        await addCol.ExecuteNonQueryAsync();
-    }
-
-    // Mark the first enabled profile as active if none are flagged yet.
-    await using var setActive = conn.CreateCommand();
-    setActive.Transaction = tx;
-    setActive.CommandText = """
-        UPDATE profiles
-        SET is_active = 1
-        WHERE profile_id = (
-            SELECT profile_id FROM profiles WHERE enabled = 1 ORDER BY created_utc ASC LIMIT 1
-        )
-        AND NOT EXISTS (SELECT 1 FROM profiles WHERE is_active = 1)
-        """;
-    await setActive.ExecuteNonQueryAsync();
-
-    await using var checkIdx = conn.CreateCommand();
-    checkIdx.Transaction = tx;
-    checkIdx.CommandText =
-        "SELECT COUNT(*) FROM sqlite_master WHERE type='index' AND name='idx_profiles_is_active'";
-    var idxExists = (long)(await checkIdx.ExecuteScalarAsync())! > 0;
-
-    if (!idxExists)
-    {
-        await using var createIdx = conn.CreateCommand();
-        createIdx.Transaction = tx;
-        createIdx.CommandText =
-            "CREATE UNIQUE INDEX \"idx_profiles_is_active\" ON \"profiles\" (\"is_active\") " +
-            "WHERE is_active = 1";
-        await createIdx.ExecuteNonQueryAsync();
-    }
-
-    await using var ins = conn.CreateCommand();
-    ins.Transaction = tx;
-    ins.CommandText =
-        "INSERT OR IGNORE INTO \"__EFMigrationsHistory\" (\"MigrationId\", \"ProductVersion\") " +
-        "VALUES (@id, @ver)";
-    var pId = ins.CreateParameter();
-    pId.ParameterName = "@id";
-    pId.Value = migrationId;
-    ins.Parameters.Add(pId);
-    var pVer = ins.CreateParameter();
-    pVer.ParameterName = "@ver";
-    pVer.Value = "10.0.0";
-    ins.Parameters.Add(pVer);
-    await ins.ExecuteNonQueryAsync();
-
-    await tx.CommitAsync();
 }
 
 static async Task PublishStartupEventsAsync(WebApplication app, AppBuildInfo buildInfo, IReadOnlyCollection<string> appliedMigrations)
